Apply gate to proper controller

+++ charbd/app/Http/Controllers/ProductController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\ImageCreateRequest;
 use App\Http\Requests\ProductCreateRequest;
 use App\Http\Resources\ProductResource;
 use App\Models\Product;
+use Gate;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Storage;
@@ -16,16 +17,19 @@ class ProductController extends Controller
 {
     public function index(Request $request)
     {
+        Gate::authorize('view', 'products');
         return ProductResource::collection(Product::paginate());
     }
 
     public function show(int $id)
     {   
+        Gate::authorize('view', 'products');
         return new ProductResource(Product::find($id));
     }
 
     public function store(ProductCreateRequest $request)
     {
+        Gate::authorize('edit', 'products');
         $product = Product::create($request->only('title', 'description', 'price', 'image'));
         return response($product, Response::HTTP_CREATED);
     }
@@ -37,6 +41,7 @@ class ProductController extends Controller
 
     public function destroy(int $id)
     {
+        Gate::authorize('edut', 'products');
         Product::destroy($id);
         return response(null, Response::HTTP_NO_CONTENT);
     }
